fix: 연관관계 설정하면서 생긴 오류 해결

+++ src/main/java/com/campuspick/fintory/domain/portfolio/domain/entity/StockTransactions.java
@@ -29,7 +29,7 @@ public class StockTransactions {
     private int amount;
 
     @Column(name = "price_per_share", precision = 15, scale = 2)
-    private float pricePerShare;
+    private BigDecimal pricePerShare;
 
     private int quantity;
